third_party/blink/web_tests/svg/custom/svg-modify-currentTranslate-expected.txt

Tests whether changes to the 'currentTranslate' SVGSVGElement property take effect. You should see an unclipped 100x100 rect above.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS svg.currentTranslate.x is 0
PASS svg.currentTranslate.y is 0
PASS svg.currentTranslate.x is 100
PASS svg.currentTranslate.y is 50
PASS svg.currentTranslate.y is 100
PASS successfullyParsed is true

TEST COMPLETE
